refactor: move EndActiveSessionEvent to events/model
