+++ src/main/java/seedu/address/model/session/Session.java
@@ -7,7 +7,7 @@ import java.util.logging.Logger;
 
 import seedu.address.commons.core.EventsCenter;
 import seedu.address.commons.core.LogsCenter;
-import seedu.address.commons.events.ui.EndActiveSessionEvent;
+import seedu.address.commons.events.model.EndActiveSessionEvent;
 import seedu.address.commons.exceptions.IllegalValueException;
 import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.CommandResult;
